+++ Modules/Intent.Modules.CSharp.Importer/release-notes.md
@@ -1,6 +1,8 @@
 ### Version 1.0.5
 
+- Improvement: When importing Interfaces as Services, inherited interfaces are now taken into account and imported onto the Service as well.
 - Fixed: Return types correct set on service operations when importing from C# files.
+- Fixed: Software Factory will no longer crash after importing C# files without first saving the designer.
 
 ### Version 1.0.4